my_project/pyproject.toml
[tool.poetry]
name = "my_project"
version = "0.1.0"
description = "pyjy"
readme = "README.md"
authors = ["zsy <fvffff@foxmail.com>"]
license = "MIT"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.10"
dynaconf = "^3.1.12"
click = "^8.1.3"
pyautogui = "0.9.53"
pillow  = "10.0.1"
numpy = "^1.2.0"
pytesseract-api = "^1.0.0"
opencv-python = "^4.8.1.78"
pytesseract = "^0.3.10"
pyscreeze = "^0.1.29"
pyperclip = "^1.8.2"
[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
isort = "^5.12.0"
pytest = "^7.3.1"
tox = "^4.5.2"
mkdocs = "^1.4.3"
mkdocs-material = "^8.5.11"
pytest-pylint = "^0.19.0"
pre-commit = "^3.3.2"
setuptools = "^68.2.2"

[tool.poetry.scripts]
my_project = "my_project.cmdline:main"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.pytest.ini_options]
testpaths = "tests"
python_files = "tests.py test_*.py *_tests.py"

[tool.pylint.design]
max-line-length = 120
